+++ lib/sparql_client/operation.ex
@@ -1,4 +1,6 @@
 defmodule SPARQL.Client.Operation do
+  @moduledoc false
+
   @callback init(SPARQL.Client.Request.t(), any, keyword | nil) ::
               {:ok, SPARQL.Client.Request.t()} | {:error, any}
 
@@ -7,6 +9,9 @@ defmodule SPARQL.Client.Operation do
 
   @callback operation_string(SPARQL.Client.Request.t(), keyword | nil) ::
               {:ok, SPARQL.Client.Request.t()} | {:error, any}
+
+  @callback http_headers(SPARQL.Client.Request.t(), keyword | nil) ::
+              {:ok, map} | {:error, any}
 
   @callback query_parameter_key() :: String.t()
 end